:recycle: minor refactor for checkout lib
