+++ lib/recart-nacelle-plugin.js
@@ -1,4 +1,4 @@
-const getMetafieldsForCheckout = require('./getMetafieldsForCheckout')
+const { getMetafieldsForCheckout } = require('./checkout')
 const { saveCart } = require('./cart')
 
 const delay = async milliseconds => {
